cmd's bufflen and data_len overflow cleanups
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@4517 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/include/scst.h
@@ -2041,7 +2041,7 @@ struct scst_cmd {
 	 * VERIFY, which transfer different amount of data (if any), than
 	 * processed.
 	 */
-	int data_len;
+	int64_t data_len;
 
 	/* Completion routine */
 	void (*scst_cmd_done) (struct scst_cmd *cmd, int next_state,
@@ -3161,7 +3161,7 @@ static inline int scst_cmd_get_bufflen(struct scst_cmd *cmd)
  * Returns cmd's data_len. See the corresponding field's description in
  * struct scst_cmd above.
  */
-static inline int scst_cmd_get_data_len(struct scst_cmd *cmd)
+static inline int64_t scst_cmd_get_data_len(struct scst_cmd *cmd)
 {
 	return cmd->data_len;
 }
