+++ scst/src/dev_handlers/scst_user.c
@@ -903,8 +903,9 @@ static int dev_user_exec(struct scst_cmd *cmd)
 	TRACE_ENTRY();
 
 	TRACE_DBG("Preparing EXEC for user space (ucmd=%p, h=%d, lba %lld, "
-		"bufflen %d, data_len %d, ubuff %lx)", ucmd, ucmd->h,
-		(long long)cmd->lba, cmd->bufflen, cmd->data_len, ucmd->ubuff);
+		"bufflen %d, data_len %lld, ubuff %lx)", ucmd, ucmd->h,
+		(long long)cmd->lba, cmd->bufflen, (long long)cmd->data_len,
+		ucmd->ubuff);
 
 	if (cmd->data_direction & SCST_DATA_WRITE)
 		dev_user_flush_dcache(ucmd);
@@ -1330,9 +1331,9 @@ static int dev_user_process_reply_parse(struct scst_user_cmd *ucmd,
 		goto out_inval;
 
 	TRACE_DBG("ucmd %p, queue_type %x, data_direction, %x, lba %lld, "
-		"bufflen %d, data_len %d, pbuf %llx, cdb_len %d, op_flags %x",
+		"bufflen %d, data_len %lld, pbuf %llx, cdb_len %d, op_flags %x",
 		ucmd, preply->queue_type, preply->data_direction,
-		(long long)preply->lba, preply->bufflen, preply->data_len,
+		(long long)preply->lba, preply->bufflen, (long long)preply->data_len,
 		reply->alloc_reply.pbuf, preply->cdb_len, preply->op_flags);
 
 	cmd->queue_type = preply->queue_type;
